+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## 0.2.0
+* expanded `tabulate` and `layout` signatures so they accept anything with trait `ToString`
+* added the `macerate` method to facilitate adding color
 ## 0.1.1
 * added this file
 * fixed a typo in the documentation